requirements.txt
asttokens==2.4.1
bcrypt==4.1.2
blinker==1.7.0
boto3==1.34.31
botocore==1.34.31
click==8.1.7
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-DebugToolbar @ git+https://github.com/pallets-eco/flask-debugtoolbar@9b63ad1837458f14597b87ad266da3d38835071f
flask-marshmallow==1.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
ipython==8.20.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.4
marshmallow==3.20.2
marshmallow-sqlalchemy==1.0.0
matplotlib-inline==0.1.6
packaging==23.2
parso==0.8.3
pexpect==4.9.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
s3transfer==0.10.0
six==1.16.0
SQLAlchemy==2.0.25
stack-data==0.6.3
traitlets==5.14.1
typing_extensions==4.9.0
urllib3==2.0.7
wcwidth==0.2.13
Werkzeug==2.3.8
WTForms==3.1.2
